@@ -1,9 +1,9 @@
 [package]
 name = "czkawka_cli"
-version = "8.0.0"
+version = "9.0.0"
 authors = ["Rafał Mikrut <mikrutrafal@protonmail.com>"]
-edition = "2021"
-rust-version = "1.79.0"
+edition = "2024"
+rust-version = "1.85.0"
 description = "CLI frontend of Czkawka"
 license = "MIT"
 homepage = "https://github.com/qarmin/czkawka"
@@ -13,18 +13,18 @@ repository = "https://github.com/qarmin/czkawka"
 clap = { version = "4.5", features = ["derive"] }
 
 # For enum types
-image_hasher = "2.0"
+image_hasher = "3.0"
 
 log = "0.4.22"
-handsome_logger = "0.8"
-fun_time = { version = "0.3", features = ["log"] }
-czkawka_core = { path = "../czkawka_core", version = "8.0.0", features = [] }
+czkawka_core = { path = "../czkawka_core", version = "9.0.0", features = [] }
 indicatif = "0.17"
 crossbeam-channel = { version = "0.5", features = [] }
 ctrlc = { version = "3.4", features = ["termination"] }
+humansize = "2.1"
 
 [features]
-default = []
+default = ["fast_image_resize"]
 heif = ["czkawka_core/heif"]
 libraw = ["czkawka_core/libraw"]
 libavif = ["czkawka_core/libavif"]
+fast_image_resize = ["czkawka_core/fast_image_resize"]
