@@ -1,9 +1,9 @@
 [package]
 name = "czkawka_core"
-version = "8.0.0"
+version = "9.0.0"
 authors = ["Rafał Mikrut <mikrutrafal@protonmail.com>"]
-edition = "2021"
-rust-version = "1.79.0"
+edition = "2024"
+rust-version = "1.85.0"
 description = "Core of Czkawka app"
 license = "MIT"
 homepage = "https://github.com/qarmin/czkawka"
@@ -19,14 +19,14 @@ crossbeam-channel = "0.5"
 directories-next = "2.0"
 
 # Needed by similar images
-image_hasher = "2.0"
+image_hasher = "3.0"
 bk-tree = "0.5"
-image = { version = "0.25", default-features = false, features = ["bmp", "dds", "exr", "ff", "gif", "hdr", "ico", "jpeg", "png", "pnm", "qoi", "tga", "tiff", "webp"] }
-hamming = "0.1"
+image = { version = "0.25", default-features = false, features = ["bmp", "dds", "exr", "ff", "gif", "hdr", "ico", "jpeg", "png", "pnm", "qoi", "tga", "tiff", "webp", "rayon"] }
+hamming-bitwise-fast = "1.0"
 
 # Needed by same music
 bitflags = "2.6"
-lofty = "0.21"
+lofty = "0.22"
 
 # Needed by broken files
 zip = { version = "2.2", features = ["aes-crypto", "bzip2", "deflate", "time"], default-features = false }
@@ -34,7 +34,7 @@ audio_checker = "0.1"
 pdf = "0.9"
 
 # Needed by audio similarity feature
-rusty-chromaprint = "0.2"
+rusty-chromaprint = "0.3"
 symphonia = { version = "0.5", features = ["all"] }
 
 # Hashes for duplicate files
@@ -45,12 +45,12 @@ xxhash-rust = { version = "0.8", features = ["xxh3"] }
 tempfile = "3.13"
 
 # Video Duplicates
-vid_dup_finder_lib = "0.1"
-ffmpeg_cmdline_utils = "0.1"
+vid_dup_finder_lib = "0.2"
+ffmpeg_cmdline_utils = "0.3"
 
 # Saving/Loading Cache
 serde = "1.0"
-bincode = "1.3"
+bincode = "<2.0" # TODO - wait for bugfix releases for 2.0 version - this may be a little unstable now, but entire cache system depends on it
 serde_json = "1.0"
 
 # Language
@@ -60,31 +60,32 @@ rust-embed = { version = "8.5", features = ["debug-embed"] }
 once_cell = "1.20"
 
 # Raw image files
-#rawler = "0.6"
-#imagepipe = { path = "/home/rafal/test/imagepipe" }
 rawloader = "0.37"
 imagepipe = "0.5"
 libraw-rs = { version = "0.0.4", optional = true }
-jxl-oxide = { version = "0.9.0", default-features = false }
+jxl-oxide = { version = "0.11.0", features = ["image"] }
 
 # Checking for invalid extensions
 mime_guess = "2.0"
-infer = "0.16"
+infer = "0.19"
 
 # Heif/Heic
 libheif-rs = { version = "=0.18.0", optional = true } # Do not upgrade now, since Ubuntu 22.04 not works with newer version
 libheif-sys = { version = "=1.14.2", optional = true } # 1.14.3 brake compilation on Ubuntu 22.04, so pin it to this version
 anyhow = { version = "1.0.89" }
 
+nom-exif = "2.1.0"
 
 state = "0.6"
 trash = "5.1"
 
 os_info = { version = "3", default-features = false }
 log = "0.4.22"
-handsome_logger = "0.8"
+handsome_logger = "0.9"
 fun_time = { version = "0.3", features = ["log"] }
-itertools = "0.13"
+itertools = "0.14"
+static_assertions = "1.1.0"
+
 
 # Don't update anymore! This crate has a bug. I've submitted a patch upstream, but the change is breaking. The current code relies on the bug to work correctly!
 # Warning by CalunVier 2024.7.15
@@ -94,8 +95,16 @@ file-id = "=0.2.1"
 [build-dependencies]
 rustc_version = "0.4"
 
+[dev-dependencies]
+criterion = "0.5"
+
+[[bench]]
+name = "hash_calculation_benchmark"
+harness = false
+
 [features]
-default = []
+default = ["fast_image_resize"]
 heif = ["dep:libheif-rs", "dep:libheif-sys"]
 libraw = ["dep:libraw-rs"]
 libavif = ["image/avif-native", "image/avif"]
+fast_image_resize = ["image_hasher/fast_resize_unstable"]
