@@ -1,9 +1,9 @@
 [package]
 name = "krokiet"
-version = "8.0.0"
+version = "9.0.0"
 authors = ["Rafał Mikrut <mikrutrafal@protonmail.com>"]
-edition = "2021"
-rust-version = "1.79.0"
+edition = "2024"
+rust-version = "1.85.0"
 description = "Slint frontend of Czkawka Core"
 license = "GPL-3.0-only"
 homepage = "https://github.com/qarmin/czkawka"
@@ -11,12 +11,10 @@ repository = "https://github.com/qarmin/czkawka"
 build = "build.rs"
 
 [dependencies]
-rand = "0.8"
-czkawka_core = { version = "8.0.0", path = "../czkawka_core" }
+czkawka_core = { version = "9.0.0", path = "../czkawka_core" }
 chrono = "0.4.38"
 open = "5.3"
 crossbeam-channel = "0.5"
-handsome_logger = "0.8"
 rfd = { version = "0.15", default-features = false, features = ["xdg-portal", "async-std"] }
 home = "0.5"
 log = "0.4.22"
@@ -24,8 +22,7 @@ serde = "1.0"
 serde_json = "1.0"
 humansize = "2.1"
 image = "0.25"
-directories-next = "2.0"
-image_hasher = "2.0"
+image_hasher = "3.0"
 rayon = "1.10"
 fs_extra = "1.3" # TODO replace with less buggy library
 trash = "5.1"
@@ -39,7 +36,7 @@ once_cell = "1.20"
 # Try to use only needed features from https://github.com/slint-ui/slint/blob/master/api/rs/slint/Cargo.toml#L23-L31
 #slint = { path = "/home/rafal/test/slint/api/rs/slint/", default-features = false, features = ["std",
 #slint = { git = "https://github.com/slint-ui/slint.git", default-features = false, features = [
-slint = { version = "1.8", default-features = false, features = [
+slint = { version = "1.10", default-features = false, features = [
     "std",
     "backend-winit",
     "compat-1-2"
@@ -63,3 +60,4 @@ winit_software = ["slint/renderer-winit-software"]
 heif = ["czkawka_core/heif"]
 libraw = ["czkawka_core/libraw"]
 libavif = ["czkawka_core/libavif"]
+fast_image_resize = ["czkawka_core/fast_image_resize"]
