@@ -2,27 +2,29 @@
 #![warn(clippy::unwrap_used)]
 
 use std::path::PathBuf;
+use std::sync::Arc;
+use std::sync::atomic::AtomicBool;
 use std::thread;
 
 use clap::Parser;
 use commands::Commands;
-use crossbeam_channel::{bounded, unbounded, Receiver, Sender};
-use czkawka_core::bad_extensions::{BadExtensions, BadExtensionsParameters};
-use czkawka_core::big_file::{BigFile, BigFileParameters, SearchMode};
-use czkawka_core::broken_files::{BrokenFiles, BrokenFilesParameters, CheckedTypes};
-use czkawka_core::common::{print_version_mode, set_number_of_threads, setup_logger, DEFAULT_THREAD_SIZE};
+use crossbeam_channel::{Receiver, Sender, unbounded};
+use czkawka_core::common::{DEFAULT_THREAD_SIZE, print_version_mode, set_config_cache_path, set_number_of_threads, setup_logger};
 use czkawka_core::common_tool::{CommonData, DeleteMethod};
 #[allow(unused_imports)] // It is used in release for print_results_to_output().
 use czkawka_core::common_traits::*;
-use czkawka_core::duplicate::{DuplicateFinder, DuplicateFinderParameters};
-use czkawka_core::empty_files::EmptyFiles;
-use czkawka_core::empty_folder::EmptyFolder;
-use czkawka_core::invalid_symlinks::InvalidSymlinks;
 use czkawka_core::progress_data::ProgressData;
-use czkawka_core::same_music::{SameMusic, SameMusicParameters};
-use czkawka_core::similar_images::{return_similarity_from_similarity_preset, test_image_conversion_speed, SimilarImages, SimilarImagesParameters};
-use czkawka_core::similar_videos::{SimilarVideos, SimilarVideosParameters};
-use czkawka_core::temporary::Temporary;
+use czkawka_core::tools::bad_extensions::{BadExtensions, BadExtensionsParameters};
+use czkawka_core::tools::big_file::{BigFile, BigFileParameters, SearchMode};
+use czkawka_core::tools::broken_files::{BrokenFiles, BrokenFilesParameters, CheckedTypes};
+use czkawka_core::tools::duplicate::{DuplicateFinder, DuplicateFinderParameters};
+use czkawka_core::tools::empty_files::EmptyFiles;
+use czkawka_core::tools::empty_folder::EmptyFolder;
+use czkawka_core::tools::invalid_symlinks::InvalidSymlinks;
+use czkawka_core::tools::same_music::{SameMusic, SameMusicParameters};
+use czkawka_core::tools::similar_images::{SimilarImages, SimilarImagesParameters, return_similarity_from_similarity_preset};
+use czkawka_core::tools::similar_videos::{SimilarVideos, SimilarVideosParameters};
+use czkawka_core::tools::temporary::Temporary;
 use log::error;
 
 use crate::commands::{
@@ -38,46 +40,62 @@ fn main() {
     let command = Args::parse().command;
 
     setup_logger(true);
-    print_version_mode();
+    print_version_mode("Czkawka cli");
+    set_config_cache_path("Czkawka", "Czkawka");
 
     if cfg!(debug_assertions) {
         println!("{command:?}");
     }
 
     let (progress_sender, progress_receiver): (Sender<ProgressData>, Receiver<ProgressData>) = unbounded();
-    let (stop_sender, stop_receiver): (Sender<()>, Receiver<()>) = bounded(1);
+    let stop_flag = Arc::new(AtomicBool::new(false));
+    let store_flag_cloned = stop_flag.clone();
+
+    let found_any_files = Arc::new(AtomicBool::new(false));
+    let found_any_files_clone = found_any_files.clone();
 
     let calculate_thread = thread::Builder::new()
         .stack_size(DEFAULT_THREAD_SIZE)
-        .spawn(move || match command {
-            Commands::Duplicates(duplicates_args) => duplicates(duplicates_args, &stop_receiver, &progress_sender),
-            Commands::EmptyFolders(empty_folders_args) => empty_folders(empty_folders_args, &stop_receiver, &progress_sender),
-            Commands::BiggestFiles(biggest_files_args) => biggest_files(biggest_files_args, &stop_receiver, &progress_sender),
-            Commands::EmptyFiles(empty_files_args) => empty_files(empty_files_args, &stop_receiver, &progress_sender),
-            Commands::Temporary(temporary_args) => temporary(temporary_args, &stop_receiver, &progress_sender),
-            Commands::SimilarImages(similar_images_args) => similar_images(similar_images_args, &stop_receiver, &progress_sender),
-            Commands::SameMusic(same_music_args) => same_music(same_music_args, &stop_receiver, &progress_sender),
-            Commands::InvalidSymlinks(invalid_symlinks_args) => invalid_symlinks(invalid_symlinks_args, &stop_receiver, &progress_sender),
-            Commands::BrokenFiles(broken_files_args) => broken_files(broken_files_args, &stop_receiver, &progress_sender),
-            Commands::SimilarVideos(similar_videos_args) => similar_videos(similar_videos_args, &stop_receiver, &progress_sender),
-            Commands::BadExtensions(bad_extensions_args) => bad_extensions(bad_extensions_args, &stop_receiver, &progress_sender),
-            Commands::Tester {} => test_image_conversion_speed(),
+        .spawn(move || {
+            let found_files = match command {
+                Commands::Duplicates(duplicates_args) => duplicates(duplicates_args, &stop_flag, &progress_sender),
+                Commands::EmptyFolders(empty_folders_args) => empty_folders(empty_folders_args, &stop_flag, &progress_sender),
+                Commands::BiggestFiles(biggest_files_args) => biggest_files(biggest_files_args, &stop_flag, &progress_sender),
+                Commands::EmptyFiles(empty_files_args) => empty_files(empty_files_args, &stop_flag, &progress_sender),
+                Commands::Temporary(temporary_args) => temporary(temporary_args, &stop_flag, &progress_sender),
+                Commands::SimilarImages(similar_images_args) => similar_images(similar_images_args, &stop_flag, &progress_sender),
+                Commands::SameMusic(same_music_args) => same_music(same_music_args, &stop_flag, &progress_sender),
+                Commands::InvalidSymlinks(invalid_symlinks_args) => invalid_symlinks(invalid_symlinks_args, &stop_flag, &progress_sender),
+                Commands::BrokenFiles(broken_files_args) => broken_files(broken_files_args, &stop_flag, &progress_sender),
+                Commands::SimilarVideos(similar_videos_args) => similar_videos(similar_videos_args, &stop_flag, &progress_sender),
+                Commands::BadExtensions(bad_extensions_args) => bad_extensions(bad_extensions_args, &stop_flag, &progress_sender),
+            };
+
+            found_any_files_clone.store(found_files, std::sync::atomic::Ordering::SeqCst);
         })
         .expect("Failed to spawn calculation thread");
+
     ctrlc::set_handler(move || {
+        if store_flag_cloned.load(std::sync::atomic::Ordering::SeqCst) {
+            return;
+        }
         println!("Get Ctrl+C signal, stopping...");
-        if let Err(e) = stop_sender.send(()) {
-            eprintln!("Failed to send stop signal {e}(it is possible that the program is already stopped)");
-        };
+        store_flag_cloned.store(true, std::sync::atomic::Ordering::SeqCst);
     })
     .expect("Error setting Ctrl-C handler");
 
     connect_progress(&progress_receiver);
 
     calculate_thread.join().expect("Failed to join calculation thread");
+
+    if found_any_files.load(std::sync::atomic::Ordering::SeqCst) {
+        std::process::exit(11);
+    } else {
+        std::process::exit(0);
+    }
 }
 
-fn duplicates(duplicates: DuplicatesArgs, stop_receiver: &Receiver<()>, progress_sender: &Sender<ProgressData>) {
+fn duplicates(duplicates: DuplicatesArgs, stop_flag: &Arc<AtomicBool>, progress_sender: &Sender<ProgressData>) -> bool {
     let DuplicatesArgs {
         common_cli_items,
         reference_directories,
@@ -111,12 +129,18 @@ fn duplicates(duplicates: DuplicatesArgs, stop_receiver: &Receiver<()>, progress
     item.set_delete_method(delete_method.delete_method);
     item.set_dry_run(dry_run.dry_run);
 
-    item.find_duplicates(Some(stop_receiver), Some(progress_sender));
+    item.find_duplicates(Some(stop_flag), Some(progress_sender));
 
-    save_and_print_results(&mut item, &common_cli_items);
+    save_and_print_results(&item, &common_cli_items);
+
+    !common_cli_items.ignore_error_code_on_found
+        && (item.get_information().number_of_duplicated_files_by_hash > 0
+            || item.get_information().number_of_duplicated_files_by_name > 0
+            || item.get_information().number_of_duplicated_files_by_size > 0
+            || item.get_information().number_of_duplicated_files_by_size_name > 0)
 }
 
-fn empty_folders(empty_folders: EmptyFoldersArgs, stop_receiver: &Receiver<()>, progress_sender: &Sender<ProgressData>) {
+fn empty_folders(empty_folders: EmptyFoldersArgs, stop_flag: &Arc<AtomicBool>, progress_sender: &Sender<ProgressData>) -> bool {
     let EmptyFoldersArgs { common_cli_items, delete_folders } = empty_folders;
 
     let mut item = EmptyFolder::new();
@@ -126,12 +150,14 @@ fn empty_folders(empty_folders: EmptyFoldersArgs, stop_receiver: &Receiver<()>, 
         item.set_delete_method(DeleteMethod::Delete);
     }
 
-    item.find_empty_folders(Some(stop_receiver), Some(progress_sender));
+    item.find_empty_folders(Some(stop_flag), Some(progress_sender));
 
-    save_and_print_results(&mut item, &common_cli_items);
+    save_and_print_results(&item, &common_cli_items);
+
+    !common_cli_items.ignore_error_code_on_found && item.get_information().number_of_empty_folders > 0
 }
 
-fn biggest_files(biggest_files: BiggestFilesArgs, stop_receiver: &Receiver<()>, progress_sender: &Sender<ProgressData>) {
+fn biggest_files(biggest_files: BiggestFilesArgs, stop_flag: &Arc<AtomicBool>, progress_sender: &Sender<ProgressData>) -> bool {
     let BiggestFilesArgs {
         common_cli_items,
         number_of_files,
@@ -148,12 +174,14 @@ fn biggest_files(biggest_files: BiggestFilesArgs, stop_receiver: &Receiver<()>, 
         item.set_delete_method(DeleteMethod::Delete);
     }
 
-    item.find_big_files(Some(stop_receiver), Some(progress_sender));
+    item.find_big_files(Some(stop_flag), Some(progress_sender));
 
-    save_and_print_results(&mut item, &common_cli_items);
+    save_and_print_results(&item, &common_cli_items);
+
+    !common_cli_items.ignore_error_code_on_found && item.get_information().number_of_real_files > 0
 }
 
-fn empty_files(empty_files: EmptyFilesArgs, stop_receiver: &Receiver<()>, progress_sender: &Sender<ProgressData>) {
+fn empty_files(empty_files: EmptyFilesArgs, stop_flag: &Arc<AtomicBool>, progress_sender: &Sender<ProgressData>) -> bool {
     let EmptyFilesArgs { common_cli_items, delete_files } = empty_files;
 
     let mut item = EmptyFiles::new();
@@ -163,12 +191,14 @@ fn empty_files(empty_files: EmptyFilesArgs, stop_receiver: &Receiver<()>, progre
         item.set_delete_method(DeleteMethod::Delete);
     }
 
-    item.find_empty_files(Some(stop_receiver), Some(progress_sender));
+    item.find_empty_files(Some(stop_flag), Some(progress_sender));
 
-    save_and_print_results(&mut item, &common_cli_items);
+    save_and_print_results(&item, &common_cli_items);
+
+    !common_cli_items.ignore_error_code_on_found && item.get_information().number_of_empty_files > 0
 }
 
-fn temporary(temporary: TemporaryArgs, stop_receiver: &Receiver<()>, progress_sender: &Sender<ProgressData>) {
+fn temporary(temporary: TemporaryArgs, stop_flag: &Arc<AtomicBool>, progress_sender: &Sender<ProgressData>) -> bool {
     let TemporaryArgs { common_cli_items, delete_files } = temporary;
 
     let mut item = Temporary::new();
@@ -178,12 +208,14 @@ fn temporary(temporary: TemporaryArgs, stop_receiver: &Receiver<()>, progress_se
         item.set_delete_method(DeleteMethod::Delete);
     }
 
-    item.find_temporary_files(Some(stop_receiver), Some(progress_sender));
+    item.find_temporary_files(Some(stop_flag), Some(progress_sender));
 
-    save_and_print_results(&mut item, &common_cli_items);
+    save_and_print_results(&item, &common_cli_items);
+
+    !common_cli_items.ignore_error_code_on_found && item.get_information().number_of_temporary_files > 0
 }
 
-fn similar_images(similar_images: SimilarImagesArgs, stop_receiver: &Receiver<()>, progress_sender: &Sender<ProgressData>) {
+fn similar_images(similar_images: SimilarImagesArgs, stop_flag: &Arc<AtomicBool>, progress_sender: &Sender<ProgressData>) -> bool {
     let SimilarImagesArgs {
         common_cli_items,
         reference_directories,
@@ -216,12 +248,14 @@ fn similar_images(similar_images: SimilarImagesArgs, stop_receiver: &Receiver<()
     item.set_delete_method(delete_method.delete_method);
     item.set_dry_run(dry_run.dry_run);
 
-    item.find_similar_images(Some(stop_receiver), Some(progress_sender));
+    item.find_similar_images(Some(stop_flag), Some(progress_sender));
 
-    save_and_print_results(&mut item, &common_cli_items);
+    save_and_print_results(&item, &common_cli_items);
+
+    !common_cli_items.ignore_error_code_on_found && item.get_information().number_of_duplicates > 0
 }
 
-fn same_music(same_music: SameMusicArgs, stop_receiver: &Receiver<()>, progress_sender: &Sender<ProgressData>) {
+fn same_music(same_music: SameMusicArgs, stop_flag: &Arc<AtomicBool>, progress_sender: &Sender<ProgressData>) -> bool {
     let SameMusicArgs {
         common_cli_items,
         reference_directories,
@@ -253,12 +287,14 @@ fn same_music(same_music: SameMusicArgs, stop_receiver: &Receiver<()>, progress_
     item.set_delete_method(delete_method.delete_method);
     item.set_dry_run(dry_run.dry_run);
 
-    item.find_same_music(Some(stop_receiver), Some(progress_sender));
+    item.find_same_music(Some(stop_flag), Some(progress_sender));
 
-    save_and_print_results(&mut item, &common_cli_items);
+    save_and_print_results(&item, &common_cli_items);
+
+    !common_cli_items.ignore_error_code_on_found && item.get_information().number_of_duplicates > 0
 }
 
-fn invalid_symlinks(invalid_symlinks: InvalidSymlinksArgs, stop_receiver: &Receiver<()>, progress_sender: &Sender<ProgressData>) {
+fn invalid_symlinks(invalid_symlinks: InvalidSymlinksArgs, stop_flag: &Arc<AtomicBool>, progress_sender: &Sender<ProgressData>) -> bool {
     let InvalidSymlinksArgs { common_cli_items, delete_files } = invalid_symlinks;
 
     let mut item = InvalidSymlinks::new();
@@ -268,12 +304,14 @@ fn invalid_symlinks(invalid_symlinks: InvalidSymlinksArgs, stop_receiver: &Recei
         item.set_delete_method(DeleteMethod::Delete);
     }
 
-    item.find_invalid_links(Some(stop_receiver), Some(progress_sender));
+    item.find_invalid_links(Some(stop_flag), Some(progress_sender));
 
-    save_and_print_results(&mut item, &common_cli_items);
+    save_and_print_results(&item, &common_cli_items);
+
+    !common_cli_items.ignore_error_code_on_found && item.get_information().number_of_invalid_symlinks > 0
 }
 
-fn broken_files(broken_files: BrokenFilesArgs, stop_receiver: &Receiver<()>, progress_sender: &Sender<ProgressData>) {
+fn broken_files(broken_files: BrokenFilesArgs, stop_flag: &Arc<AtomicBool>, progress_sender: &Sender<ProgressData>) -> bool {
     let BrokenFilesArgs {
         common_cli_items,
         delete_files,
@@ -292,12 +330,14 @@ fn broken_files(broken_files: BrokenFilesArgs, stop_receiver: &Receiver<()>, pro
         item.set_delete_method(DeleteMethod::Delete);
     }
 
-    item.find_broken_files(Some(stop_receiver), Some(progress_sender));
+    item.find_broken_files(Some(stop_flag), Some(progress_sender));
 
-    save_and_print_results(&mut item, &common_cli_items);
+    save_and_print_results(&item, &common_cli_items);
+
+    !common_cli_items.ignore_error_code_on_found && item.get_information().number_of_broken_files > 0
 }
 
-fn similar_videos(similar_videos: SimilarVideosArgs, stop_receiver: &Receiver<()>, progress_sender: &Sender<ProgressData>) {
+fn similar_videos(similar_videos: SimilarVideosArgs, stop_flag: &Arc<AtomicBool>, progress_sender: &Sender<ProgressData>) -> bool {
     let SimilarVideosArgs {
         reference_directories,
         common_cli_items,
@@ -319,12 +359,14 @@ fn similar_videos(similar_videos: SimilarVideosArgs, stop_receiver: &Receiver<()
     item.set_delete_method(delete_method.delete_method);
     item.set_dry_run(dry_run.dry_run);
 
-    item.find_similar_videos(Some(stop_receiver), Some(progress_sender));
+    item.find_similar_videos(Some(stop_flag), Some(progress_sender));
 
-    save_and_print_results(&mut item, &common_cli_items);
+    save_and_print_results(&item, &common_cli_items);
+
+    !common_cli_items.ignore_error_code_on_found && item.get_information().number_of_duplicates > 0
 }
 
-fn bad_extensions(bad_extensions: BadExtensionsArgs, stop_receiver: &Receiver<()>, progress_sender: &Sender<ProgressData>) {
+fn bad_extensions(bad_extensions: BadExtensionsArgs, stop_flag: &Arc<AtomicBool>, progress_sender: &Sender<ProgressData>) -> bool {
     let BadExtensionsArgs { common_cli_items } = bad_extensions;
 
     let params = BadExtensionsParameters::new();
@@ -332,12 +374,14 @@ fn bad_extensions(bad_extensions: BadExtensionsArgs, stop_receiver: &Receiver<()
 
     set_common_settings(&mut item, &common_cli_items, None);
 
-    item.find_bad_extensions_files(Some(stop_receiver), Some(progress_sender));
+    item.find_bad_extensions_files(Some(stop_flag), Some(progress_sender));
 
-    save_and_print_results(&mut item, &common_cli_items);
+    save_and_print_results(&item, &common_cli_items);
+
+    !common_cli_items.ignore_error_code_on_found && item.get_information().number_of_files_with_bad_extension > 0
 }
 
-fn save_and_print_results<T: CommonData + PrintResults>(component: &mut T, common_cli_items: &CommonCliItems) {
+fn save_and_print_results<T: CommonData + PrintResults>(component: &T, common_cli_items: &CommonCliItems) {
     if let Some(file_name) = common_cli_items.file_to_save.file_name() {
         if let Err(e) = component.print_results_to_file(file_name) {
             error!("Failed to save results to file {e}");
@@ -354,9 +398,13 @@ fn save_and_print_results<T: CommonData + PrintResults>(component: &mut T, commo
         }
     }
 
-    component.print_results_to_output();
+    if !common_cli_items.do_not_print.do_not_print_results {
+        component.print_results_to_output();
+    }
 
-    component.get_text_messages().print_messages();
+    if !common_cli_items.do_not_print.do_not_print_messages {
+        component.get_text_messages().print_messages();
+    }
 }
 
 fn set_common_settings<T>(component: &mut T, common_cli_items: &CommonCliItems, reference_directories: Option<&Vec<PathBuf>>)
@@ -378,4 +426,5 @@ where
     #[cfg(target_family = "unix")]
     component.set_exclude_other_filesystems(common_cli_items.exclude_other_filesystems);
     component.set_allowed_extensions(common_cli_items.allowed_extensions.clone().join(","));
+    component.set_use_cache(!common_cli_items.disable_cache);
 }
