@@ -111,8 +111,11 @@ impl Extensions {
         let mut new_extensions = HashSet::new();
         for extension in file_extensions {
             let extension_without_dot = extension.trim_start_matches('.');
-            new_extensions.insert(extension_without_dot.to_string());
+            if !self.allowed_extensions_hashset.contains(extension_without_dot) {
+                new_extensions.insert(extension_without_dot.to_string());
+            }
         }
+        self.allowed_extensions_hashset = new_extensions;
     }
 
     pub fn set_and_validate_allowed_extensions(&mut self, file_extensions: &[&str]) {
