@@ -1,14 +1,15 @@
 use std::collections::{BTreeMap, BTreeSet, HashMap, HashSet};
 use std::io::Write;
 use std::path::{Path, PathBuf};
-use std::sync::atomic::Ordering;
-use std::time::SystemTime;
+use std::sync::Arc;
+use std::sync::atomic::{AtomicBool, Ordering};
 use std::{mem, panic};
 
 use bk_tree::BKTree;
-use crossbeam_channel::{Receiver, Sender};
+use crossbeam_channel::Sender;
 use fun_time::fun_time;
-use humansize::{format_size, BINARY};
+use hamming_bitwise_fast::hamming_bitwise_fast;
+use humansize::{BINARY, format_size};
 use image::GenericImageView;
 use image_hasher::{FilterType, HashAlg, HasherConfig};
 use log::debug;
@@ -16,11 +17,11 @@ use rayon::prelude::*;
 use serde::{Deserialize, Serialize};
 
 use crate::common::{
-    check_if_stop_received, delete_files_custom, prepare_thread_handler_common, send_info_and_wait_for_ending_all_threads, HEIC_EXTENSIONS, IMAGE_RS_SIMILAR_IMAGES_EXTENSIONS,
-    JXL_IMAGE_EXTENSIONS, RAW_IMAGE_EXTENSIONS,
+    HEIC_EXTENSIONS, IMAGE_RS_SIMILAR_IMAGES_EXTENSIONS, JXL_IMAGE_EXTENSIONS, RAW_IMAGE_EXTENSIONS, WorkContinueStatus, check_if_stop_received, delete_files_custom,
+    prepare_thread_handler_common, send_info_and_wait_for_ending_all_threads,
 };
 use crate::common_cache::{extract_loaded_cache, get_similar_images_cache_file, load_cache_from_file_generalized_by_path, save_cache_to_file_generalized};
-use crate::common_dir_traversal::{inode, take_1_per_inode, DirTraversalBuilder, DirTraversalResult, FileEntry, ToolType};
+use crate::common_dir_traversal::{DirTraversalBuilder, DirTraversalResult, FileEntry, ToolType, inode, take_1_per_inode};
 use crate::common_image::get_dynamic_image_from_path;
 use crate::common_tool::{CommonData, CommonToolData, DeleteMethod};
 use crate::common_traits::{DebugPrint, PrintResults, ResultEntry};
@@ -92,7 +93,7 @@ struct Hamming;
 
 impl bk_tree::Metric<ImHash> for Hamming {
     fn distance(&self, a: &ImHash, b: &ImHash) -> u32 {
-        hamming::distance_fast(a, b).expect("Calculating hamming distance, cannot fail") as u32
+        hamming_bitwise_fast(a, b)
     }
 
     fn threshold_distance(&self, a: &ImHash, b: &ImHash, _threshold: u32) -> Option<u32> {
@@ -156,18 +157,18 @@ impl SimilarImages {
     }
 
     #[fun_time(message = "find_similar_images", level = "info")]
-    pub fn find_similar_images(&mut self, stop_receiver: Option<&Receiver<()>>, progress_sender: Option<&Sender<ProgressData>>) {
+    pub fn find_similar_images(&mut self, stop_flag: Option<&Arc<AtomicBool>>, progress_sender: Option<&Sender<ProgressData>>) {
         self.prepare_items();
         self.common_data.use_reference_folders = !self.common_data.directories.reference_directories.is_empty();
-        if !self.check_for_similar_images(stop_receiver, progress_sender) {
+        if self.check_for_similar_images(stop_flag, progress_sender) == WorkContinueStatus::Stop {
             self.common_data.stopped_search = true;
             return;
         }
-        if !self.hash_images(stop_receiver, progress_sender) {
+        if self.hash_images(stop_flag, progress_sender) == WorkContinueStatus::Stop {
             self.common_data.stopped_search = true;
             return;
         }
-        if !self.find_similar_hashes(stop_receiver, progress_sender) {
+        if self.find_similar_hashes(stop_flag, progress_sender) == WorkContinueStatus::Stop {
             self.common_data.stopped_search = true;
             return;
         }
@@ -176,7 +177,7 @@ impl SimilarImages {
     }
 
     #[fun_time(message = "check_for_similar_images", level = "debug")]
-    fn check_for_similar_images(&mut self, stop_receiver: Option<&Receiver<()>>, progress_sender: Option<&Sender<ProgressData>>) -> bool {
+    fn check_for_similar_images(&mut self, stop_flag: Option<&Arc<AtomicBool>>, progress_sender: Option<&Sender<ProgressData>>) -> WorkContinueStatus {
         if cfg!(feature = "heif") {
             self.common_data
                 .extensions
@@ -188,12 +189,12 @@ impl SimilarImages {
         }
 
         if !self.common_data.extensions.set_any_extensions() {
-            return true;
+            return WorkContinueStatus::Continue;
         }
 
         let result = DirTraversalBuilder::new()
             .group_by(inode)
-            .stop_receiver(stop_receiver)
+            .stop_flag(stop_flag)
             .progress_sender(progress_sender)
             .common_data(&self.common_data)
             .build()
@@ -202,7 +203,7 @@ impl SimilarImages {
         match result {
             DirTraversalResult::SuccessFiles { grouped_file_entries, warnings } => {
                 self.images_to_check = grouped_file_entries
-                    .into_iter()
+                    .into_par_iter()
                     .flat_map(if self.get_params().ignore_hard_links { |(_, fes)| fes } else { take_1_per_inode })
                     .map(|fe| {
                         let fe_str = fe.path.to_string_lossy().to_string();
@@ -211,12 +212,13 @@ impl SimilarImages {
                         (fe_str, image_entry)
                     })
                     .collect();
+
                 self.common_data.text_messages.warnings.extend(warnings);
                 debug!("check_files - Found {} image files.", self.images_to_check.len());
-                true
+                WorkContinueStatus::Continue
             }
 
-            DirTraversalResult::Stopped => false,
+            DirTraversalResult::Stopped => WorkContinueStatus::Stop,
         }
     }
 
@@ -265,34 +267,35 @@ impl SimilarImages {
     // - Join all hashes and save it to file
 
     #[fun_time(message = "hash_images", level = "debug")]
-    fn hash_images(&mut self, stop_receiver: Option<&Receiver<()>>, progress_sender: Option<&Sender<ProgressData>>) -> bool {
+    fn hash_images(&mut self, stop_flag: Option<&Arc<AtomicBool>>, progress_sender: Option<&Sender<ProgressData>>) -> WorkContinueStatus {
         if self.images_to_check.is_empty() {
-            return true;
+            return WorkContinueStatus::Continue;
         }
 
         let (loaded_hash_map, records_already_cached, non_cached_files_to_check) = self.hash_images_load_cache();
 
-        let (progress_thread_handle, progress_thread_run, atomic_counter, check_was_stopped) = prepare_thread_handler_common(
+        let (progress_thread_handle, progress_thread_run, items_counter, check_was_stopped, size_counter) = prepare_thread_handler_common(
             progress_sender,
             CurrentStage::SimilarImagesCalculatingHashes,
             non_cached_files_to_check.len(),
             self.get_test_type(),
+            non_cached_files_to_check.values().map(|entry| entry.size).sum(),
         );
 
         debug!("hash_images - start hashing images");
         let (mut vec_file_entry, errors): (Vec<ImagesEntry>, Vec<String>) = non_cached_files_to_check
             .into_par_iter()
-            .map(|(_s, mut file_entry)| {
-                atomic_counter.fetch_add(1, Ordering::Relaxed);
-                if check_if_stop_received(stop_receiver) {
+            .map(|(_s, file_entry)| {
+                if check_if_stop_received(stop_flag) {
                     check_was_stopped.store(true, Ordering::Relaxed);
                     return None;
                 }
-                if let Err(e) = self.collect_image_file_entry(&mut file_entry) {
-                    return Some(Err(e));
-                }
+                let size = file_entry.size;
+                let res = self.collect_image_file_entry(file_entry);
+                items_counter.fetch_add(1, Ordering::Relaxed);
+                size_counter.fetch_add(size, Ordering::Relaxed);
 
-                Some(Ok(file_entry))
+                Some(res)
             })
             .while_some()
             .partition_map(|res| match res {
@@ -322,10 +325,10 @@ impl SimilarImages {
 
         // Break if stop was clicked after saving to cache
         if check_was_stopped.load(Ordering::Relaxed) {
-            return false;
+            return WorkContinueStatus::Stop;
         }
 
-        true
+        WorkContinueStatus::Continue
     }
 
     #[fun_time(message = "save_to_cache", level = "debug")]
@@ -347,7 +350,7 @@ impl SimilarImages {
         }
     }
 
-    fn collect_image_file_entry(&self, file_entry: &mut ImagesEntry) -> Result<(), String> {
+    fn collect_image_file_entry(&self, mut file_entry: ImagesEntry) -> Result<ImagesEntry, String> {
         let img = get_dynamic_image_from_path(&file_entry.path.to_string_lossy())?;
 
         let dimensions = img.dimensions();
@@ -363,7 +366,7 @@ impl SimilarImages {
         let hash = hasher.hash_image(&img);
         file_entry.hash = hash.as_bytes().to_vec();
 
-        Ok(())
+        Ok(file_entry)
     }
 
     // Split hashes at 2 parts, base hashes and hashes to compare, 3 argument is set of hashes with multiple images
@@ -445,14 +448,14 @@ impl SimilarImages {
         all_hashed_images: &HashMap<ImHash, Vec<ImagesEntry>>,
         collected_similar_images: &mut HashMap<ImHash, Vec<ImagesEntry>>,
         progress_sender: Option<&Sender<ProgressData>>,
-        stop_receiver: Option<&Receiver<()>>,
+        stop_flag: Option<&Arc<AtomicBool>>,
         tolerance: u32,
-    ) -> bool {
+    ) -> WorkContinueStatus {
         // Don't use hashes with multiple images in bktree, because they will always be master of group and cannot be find by other hashes
         let (base_hashes, hashes_with_multiple_images) = self.split_hashes(all_hashed_images);
 
-        let (progress_thread_handle, progress_thread_run, atomic_counter, check_was_stopped) =
-            prepare_thread_handler_common(progress_sender, CurrentStage::SimilarImagesComparingHashes, base_hashes.len(), self.get_test_type());
+        let (progress_thread_handle, progress_thread_run, items_counter, check_was_stopped, _size_counter) =
+            prepare_thread_handler_common(progress_sender, CurrentStage::SimilarImagesComparingHashes, base_hashes.len(), self.get_test_type(), 0);
 
         let mut hashes_parents: HashMap<ImHash, u32> = Default::default(); // Hashes used as parent (hash, children_number_of_hash)
         let mut hashes_similarity: HashMap<ImHash, (ImHash, u32)> = Default::default(); // Hashes used as child, (parent_hash, similarity)
@@ -464,9 +467,9 @@ impl SimilarImages {
             let partial_results = chunk
                 .into_par_iter()
                 .map(|hash_to_check| {
-                    atomic_counter.fetch_add(1, Ordering::Relaxed);
+                    items_counter.fetch_add(1, Ordering::Relaxed);
 
-                    if check_if_stop_received(stop_receiver) {
+                    if check_if_stop_received(stop_flag) {
                         check_was_stopped.store(true, Ordering::Relaxed);
                         return None;
                     }
@@ -502,7 +505,7 @@ impl SimilarImages {
 
             if check_was_stopped.load(Ordering::Relaxed) {
                 send_info_and_wait_for_ending_all_threads(&progress_thread_run, progress_thread_handle);
-                return false;
+                return WorkContinueStatus::Stop;
             }
 
             self.connect_results(partial_results, &mut hashes_parents, &mut hashes_similarity, &hashes_with_multiple_images);
@@ -513,7 +516,7 @@ impl SimilarImages {
         debug_check_for_duplicated_things(self.common_data.use_reference_folders, &hashes_parents, &hashes_similarity, all_hashed_images, "LATTER");
         self.collect_hash_compare_result(hashes_parents, &hashes_with_multiple_images, all_hashed_images, collected_similar_images, hashes_similarity);
 
-        true
+        WorkContinueStatus::Continue
     }
 
     #[fun_time(message = "connect_results", level = "debug")]
@@ -581,9 +584,9 @@ impl SimilarImages {
     }
 
     #[fun_time(message = "find_similar_hashes", level = "debug")]
-    fn find_similar_hashes(&mut self, stop_receiver: Option<&Receiver<()>>, progress_sender: Option<&Sender<ProgressData>>) -> bool {
+    fn find_similar_hashes(&mut self, stop_flag: Option<&Arc<AtomicBool>>, progress_sender: Option<&Sender<ProgressData>>) -> WorkContinueStatus {
         if self.image_hashes.is_empty() {
-            return true;
+            return WorkContinueStatus::Continue;
         }
 
         let tolerance = self.get_params().similarity;
@@ -600,8 +603,9 @@ impl SimilarImages {
                     collected_similar_images.insert(hash, vec_file_entry);
                 }
             }
-        } else if !self.compare_hashes_with_non_zero_tolerance(&all_hashed_images, &mut collected_similar_images, progress_sender, stop_receiver, tolerance) {
-            return false;
+        } else if self.compare_hashes_with_non_zero_tolerance(&all_hashed_images, &mut collected_similar_images, progress_sender, stop_flag, tolerance) == WorkContinueStatus::Stop
+        {
+            return WorkContinueStatus::Stop;
         }
 
         self.verify_duplicated_items(&collected_similar_images);
@@ -630,7 +634,7 @@ impl SimilarImages {
         self.images_to_check = Default::default();
         self.bktree = BKTree::new(Hamming);
 
-        true
+        WorkContinueStatus::Continue
     }
 
     #[fun_time(message = "exclude_items_with_same_size", level = "debug")]
@@ -640,8 +644,7 @@ impl SimilarImages {
                 let mut bt_sizes: BTreeSet<u64> = Default::default();
                 let mut vec_values = Vec::new();
                 for file_entry in vec_file_entry {
-                    if !bt_sizes.contains(&file_entry.size) {
-                        bt_sizes.insert(file_entry.size);
+                    if bt_sizes.insert(file_entry.size) {
                         vec_values.push(file_entry);
                     }
                 }
@@ -867,61 +870,6 @@ pub fn convert_algorithm_to_string(hash_alg: &HashAlg) -> String {
     .to_string()
 }
 
-pub fn test_image_conversion_speed() {
-    let file_name: &str = "test.jpg";
-    let file_path = Path::new(file_name);
-    match image::open(file_path) {
-        Ok(img_open) => {
-            for alg in [
-                HashAlg::Blockhash,
-                HashAlg::Gradient,
-                HashAlg::DoubleGradient,
-                HashAlg::VertGradient,
-                HashAlg::Mean,
-                HashAlg::Median,
-            ] {
-                for filter in [
-                    FilterType::Lanczos3,
-                    FilterType::CatmullRom,
-                    FilterType::Gaussian,
-                    FilterType::Nearest,
-                    FilterType::Triangle,
-                ] {
-                    for size in [8, 16, 32, 64] {
-                        let hasher_config = HasherConfig::new().hash_alg(alg).resize_filter(filter).hash_size(size, size);
-
-                        let start = SystemTime::now();
-
-                        let hasher = hasher_config.to_hasher();
-                        let _hash = hasher.hash_image(&img_open);
-
-                        let end = SystemTime::now();
-
-                        println!(
-                            "{:?} us {:?} {:?} {}x{}",
-                            end.duration_since(start).expect("Used time backwards").as_micros(),
-                            alg,
-                            filter,
-                            size,
-                            size
-                        );
-                    }
-                }
-            }
-        }
-        Err(e) => {
-            println!(
-                "Failed to open test file {}, reason {}",
-                match file_path.canonicalize() {
-                    Ok(t) => t.to_string_lossy().to_string(),
-                    Err(_inspected) => file_name.to_string(),
-                },
-                e
-            );
-        }
-    }
-}
-
 #[allow(dead_code)]
 #[allow(unreachable_code)]
 #[allow(unused_variables)]
@@ -1025,7 +973,7 @@ mod tests {
     use image_hasher::HashAlg;
 
     use crate::common_tool::CommonData;
-    use crate::similar_images::{Hamming, ImHash, ImagesEntry, SimilarImages, SimilarImagesParameters};
+    use crate::tools::similar_images::{Hamming, ImHash, ImagesEntry, SimilarImages, SimilarImagesParameters};
 
     fn get_default_parameters() -> SimilarImagesParameters {
         SimilarImagesParameters {
