@@ -1,7 +1,7 @@
 use std::fs;
 use std::path::PathBuf;
 
-use czkawka_core::duplicate::make_hard_link;
+use czkawka_core::common::make_hard_link;
 use gtk4::prelude::*;
 use gtk4::{Align, CheckButton, Dialog, Orientation, ResponseType, TextView, TreeIter, TreePath};
 
@@ -200,7 +200,6 @@ fn hardlink_symlink(
             for file_to_hardlink in symhardlink_data.files_to_symhardlink {
                 if let Err(e) = make_hard_link(&PathBuf::from(&symhardlink_data.original_data), &PathBuf::from(&file_to_hardlink)) {
                     add_text_to_text_view(text_view_errors, format!("{} {}, reason {}", flg!("hardlink_failed"), file_to_hardlink, e).as_str());
-                    continue;
                 }
             }
         }
@@ -216,14 +215,12 @@ fn hardlink_symlink(
                 {
                     if let Err(e) = std::os::unix::fs::symlink(&symhardlink_data.original_data, &file_to_symlink) {
                         add_text_to_text_view(text_view_errors, flg!("delete_file_failed", name = file_to_symlink, reason = e.to_string()).as_str());
-                        continue;
                     };
                 }
                 #[cfg(target_family = "windows")]
                 {
                     if let Err(e) = std::os::windows::fs::symlink_file(&symhardlink_data.original_data, &file_to_symlink) {
                         add_text_to_text_view(text_view_errors, flg!("delete_file_failed", name = file_to_symlink, reason = e.to_string()).as_str());
-                        continue;
                     };
                 }
             }
