@@ -3,7 +3,7 @@ use gtk4::prelude::*;
 use crate::gui_structs::gui_data::GuiData;
 use crate::gui_structs::gui_popovers_sort::GuiSortPopovers;
 use crate::help_functions::PopoverTypes;
-use crate::notebook_enums::{to_notebook_main_enum, NotebookMainEnum};
+use crate::notebook_enums::{NotebookMainEnum, to_notebook_main_enum};
 use crate::notebook_info::NOTEBOOKS_INFO;
 
 pub fn connect_button_sort(gui_data: &GuiData) {
