@@ -25,8 +25,8 @@ pub mod tbp_flags {
 pub struct TaskbarProgress {}
 
 impl TaskbarProgress {
-    pub fn new() -> TaskbarProgress {
-        TaskbarProgress {}
+    pub fn new() -> Self {
+        Self {}
     }
 
     pub fn set_progress_state(&self, _tbp_flags: TBPFLAG) {}
@@ -37,12 +37,13 @@ impl TaskbarProgress {
 
     pub fn show(&self) {}
 
+    #[allow(clippy::needless_pass_by_ref_mut)]
     pub fn release(&mut self) {}
 }
 
 impl From<HWND> for TaskbarProgress {
     fn from(_hwnd: HWND) -> Self {
-        TaskbarProgress {}
+        Self {}
     }
 }
 
