@@ -1,7 +1,7 @@
+use crate::GuiData;
 use crate::help_functions::get_notebook_enum_from_tree_view;
 use crate::notebook_enums::to_notebook_main_enum;
 use crate::notebook_info::NOTEBOOKS_INFO;
-use crate::GuiData;
 
 pub fn validate_notebook_data(gui_data: &GuiData) {
     // Test treeviews names, each treeview should have set name same as variable name
