@@ -1,4 +1,4 @@
-use directories_next::ProjectDirs;
+use czkawka_core::common::get_config_cache_path;
 use log::error;
 use slint::ComponentHandle;
 
@@ -16,25 +16,22 @@ pub fn connect_open_items(app: &MainWindow) {
     });
 
     app.global::<Callabler>().on_open_config_folder(move || {
-        let Some(dirs) = ProjectDirs::from("pl", "Qarmin", "Krokiet") else {
+        let Some(config_cache) = get_config_cache_path() else {
             error!("Failed to open config folder");
             return;
         };
-        let config_folder = dirs.config_dir();
-        if let Err(e) = open::that(config_folder) {
-            error!("Failed to open config folder {:?}: {e}", config_folder);
+        if let Err(e) = open::that(&config_cache.config_folder) {
+            error!("Failed to open config folder \"{}\": {e}", config_cache.config_folder.to_string_lossy());
         }
     });
 
-    // Cache uses Czkawka name to easily change between apps
     app.global::<Callabler>().on_open_cache_folder(move || {
-        let Some(dirs) = ProjectDirs::from("pl", "Qarmin", "Czkawka") else {
+        let Some(config_cache) = get_config_cache_path() else {
             error!("Failed to open cache folder");
             return;
         };
-        let cache_folder = dirs.cache_dir();
-        if let Err(e) = open::that(cache_folder) {
-            error!("Failed to open cache folder {:?}: {e}", cache_folder);
+        if let Err(e) = open::that(&config_cache.cache_folder) {
+            error!("Failed to open cache folder \"{}\": {e}", config_cache.cache_folder.to_string_lossy());
         }
     });
 
