@@ -1,9 +1,10 @@
-use crossbeam_channel::Sender;
+use std::sync::Arc;
+use std::sync::atomic::AtomicBool;
 
 use crate::MainWindow;
 
-pub fn connect_stop_button(app: &MainWindow, stop_sender: Sender<()>) {
+pub fn connect_stop_button(app: &MainWindow, stop_sender: Arc<AtomicBool>) {
     app.on_scan_stopping(move || {
-        stop_sender.send(()).expect("Failed to send stop signal, no much to do");
+        stop_sender.store(true, std::sync::atomic::Ordering::Relaxed);
     });
 }
