@@ -1,28 +1,16 @@
 // Remove console window in Windows OS
 #![windows_subsystem = "windows"]
 #![allow(unknown_lints)] // May be disabled, but locally I use nightly clippy
-#![allow(clippy::comparison_chain)]
 #![allow(clippy::collapsible_if)]
 #![allow(clippy::collapsible_else_if)]
-#![allow(clippy::should_panic_without_expect)]
-#![allow(clippy::struct_field_names)] // Generated code
-#![allow(clippy::overly_complex_bool_expr)] // Generated code
-#![allow(clippy::semicolon_if_nothing_returned)] // Generated code
-#![allow(clippy::used_underscore_binding)] // Generated code
-#![allow(clippy::unreadable_literal)] // Generated code
-#![allow(clippy::float_cmp)] // Generated code
-#![allow(clippy::no_effect_underscore_binding)] // Generated code
-#![allow(clippy::uninlined_format_args)] // Generated code
-#![allow(clippy::needless_pass_by_value)] // Generated code
-#![allow(clippy::redundant_closure_for_method_calls)] // Generated code
-#![allow(clippy::items_after_statements)] // Generated code
-#![allow(clippy::match_same_arms)] // Generated code
+// #![warn(clippy::unwrap_used)] // Cannot use due unwrap used in a lot of places
 
 use std::rc::Rc;
 use std::sync::Arc;
+use std::sync::atomic::AtomicBool;
 
-use crossbeam_channel::{unbounded, Receiver, Sender};
-use czkawka_core::common::{print_version_mode, setup_logger};
+use crossbeam_channel::{Receiver, Sender, unbounded};
+use czkawka_core::common::{print_version_mode, set_config_cache_path, setup_logger};
 use czkawka_core::progress_data::ProgressData;
 use slint::VecModel;
 
@@ -64,12 +52,13 @@ mod shared_models;
 slint::include_modules!();
 fn main() {
     setup_logger(false);
-    print_version_mode();
+    print_version_mode("Krokiet");
+    set_config_cache_path("Czkawka", "Krokiet");
 
     let app = MainWindow::new().expect("Failed to create main window");
 
     let (progress_sender, progress_receiver): (Sender<ProgressData>, Receiver<ProgressData>) = unbounded();
-    let (stop_sender, stop_receiver): (Sender<()>, Receiver<()>) = unbounded();
+    let stop_flag: Arc<AtomicBool> = Arc::new(AtomicBool::new(false));
 
     zeroing_all_models(&app);
 
@@ -81,8 +70,8 @@ fn main() {
     load_settings_from_file(&app);
 
     connect_delete_button(&app);
-    connect_scan_button(&app, progress_sender, stop_receiver, Arc::clone(&shared_models));
-    connect_stop_button(&app, stop_sender);
+    connect_scan_button(&app, progress_sender, stop_flag.clone(), Arc::clone(&shared_models));
+    connect_stop_button(&app, stop_flag);
     connect_open_items(&app);
     connect_progress_gathering(&app, progress_receiver);
     connect_add_remove_directories(&app);
